Chore : 잘못된 스크립트 명 수정
