+++ Assets/Scripts/Objects/Unit/Status/StatusData.cs.meta
@@ -1,5 +1,5 @@
 fileFormatVersion: 2
-guid: 3b8760d863817814db97dd4e69858620
+guid: 66bfe240957c794439fa50e17ad4e222
 MonoImporter:
   externalObjects: {}
   serializedVersion: 2
