--- Assets/Scripts/Objects/Unit/Status/StatusName.cs
@@ -1,19 +1,0 @@
-using System.Collections;
-using System.Collections.Generic;
-using UnityEngine;
-
-public enum StatusType
-{
-	NONE,
-	
-	CURRENT_HP,
-	MAX_HP,
-	SPEED,
-	ATTACK_POINT,
-	DEFENCE_POINT,
-	CRITICAL_CHANCE,
-	CRITICAL_DAMAGE_MULTIPLIER,
-	DASH_SPEED,
-
-	MAX
-}
