+++ Assets/Scripts/Objects/Unit/Status/StatusType.cs
@@ -1,45 +1,19 @@
-using System;
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
-
-[System.Serializable]
-public class StatusData
+public enum StatusType
 {
-	public StatusData(StatusType type = StatusType.NONE, float value = .0f)
-	{
-		this.type = type;
-		this.value = value;
-	}
+	NONE,
+	
+	CURRENT_HP,
+	MAX_HP,
+	SPEED,
+	ATTACK_POINT,
+	DEFENCE_POINT,
+	CRITICAL_CHANCE,
+	CRITICAL_DAMAGE_MULTIPLIER,
+	DASH_SPEED,
 
-	public StatusData(StatusData data)
-	{
-		this.type = data.type;
-		value = data.value;
-	}
-	
-	public StatusType type;
-	[SerializeField] private float value;
-
-	public float GetValue()
-	{
-		return value;
-	}
-
-	public void SetValue(float value)
-	{
-		this.value = value;
-	}
-	
-	public void PlusValue(float value)
-	{
-		this.value += value;
-	}
-	
-	public void MinusValue(float value)
-	{
-		this.value -= value;
-	}
-	
+	MAX
 }
